fix bug

+++ db/migrate/20140516101021_create_jurnal_details.rb
@@ -5,8 +5,8 @@ class CreateJurnalDetails < ActiveRecord::Migration
       t.integer :rec_account_id
       t.integer :index_jurnal
       t.integer :pos
-      t.double :debet
-      t.double :kredit
+      t.decimal :debet
+      t.decimal :kredit
       t.timestamps
     end
   end
